Setting up game server and establishing connection with client
