+++ README.md
@@ -1,2 +1,10 @@
 # org
 Online version of the Jubal game
+
+Run this to set up scaffolding:
+
+```express --view=pug org```
+
+Then after running npm i, run this to start the server:
+
+```DEBUG=org:* npm start```
