+++ client/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "firebase": "^5.11.1",
     "jquery": "^3.4.1",
-    "socket.io": "^2.2.0"
+    "jqueryui": "^1.11.1",
+    "socket.io-client": "^2.2.0"
   }
 }
