--- client/public/game/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="user-scalable=no, initial-scale=1, maximum-scale=1, minimum-scale=1, width=device-width">
-        <title>Org: The Game</title>
-        <link rel="stylesheet" type="text/css" href="css/main.css">
-    </head>
-    <body>
-
-    <header>Org Login</header>
-
-    <div id="firebaseui-auth-container"></div>
-
-<!--    <script src="https://www.gstatic.com/firebasejs/5.2.0/firebase-app.js"></script>-->
-<!--    <script src="https://www.gstatic.com/firebasejs/5.2.0/firebase-auth.js"></script>-->
-<!--    <script src="https://www.gstatic.com/firebasejs/5.2.0/firebase-database.js"></script>-->
-<!--    <script src="https://cdn.firebase.com/libs/firebaseui/3.1.1/firebaseui.js"></script>-->
-<!--    <script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>-->
-<!--    <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.min.js" integrity="sha256-VazP97ZCwtekAsvgPBSUwPFKdrwD3unUfSGVYrahUqU=" crossorigin="anonymous"></script>-->
-    <script type="text/javascript" src="js/firebase.js"></script>
-    <script type="text/javascript" src="js/events.js"></script>
-    <script type="text/javascript" src="js/ui.js"></script>
-    <script type="text/javascript" src="js/player.js"></script>
-    <script type="text/javascript" src="js/board.js"></script>
-    <script type="text/javascript" src="js/table.js"></script>
-    <script type="text/javascript" src="js/audio.js"></script>
-    <script type="text/javascript" src="js/helpers.js"></script>
-    <script type="text/javascript" src="js/turn-controller.js"></script>
-    <script type="text/javascript" src="js/main.js"></script>
-    </body>
-</html>
