+++ client/public/gamefiles/js/helpers.js
