+++ client/public/gamefiles/js/player.js
