+++ client/public/gamefiles/js/table.js
