+++ client/public/index.html
@@ -25,7 +25,7 @@
     <div id="message">
       <h1>Org Dashboard</h1>
       <a href="tools/index.html">Tools</a>
-      <a href="game/index.html">Game</a>
+      <a href="gamefiles/index.html">Game</a>
     </div>
   </body>
 </html>
