+++ server/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "express": "^4.17.1",
     "firebase": "^5.11.1",
+    "jqueryui": "^1.11.1",
     "socket.io": "^2.2.0"
   }
 }
